handle authorization with jwt guard
