+++ app/api/.env_sample
@@ -17,4 +17,4 @@ REDIRECT_URI=
 
 # JWT
 JWT_SECRET=jwtSecuredSecret
-JWT_EXPIRES_IN=1d
+JWT_EXPIRES_IN="1d"
