+++ app/api/package.json
@@ -34,6 +34,7 @@
     "class-validator": "^0.14.0",
     "dotenv": "^16.0.3",
     "passport": "^0.6.0",
+    "passport-jwt": "^4.0.1",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.2.0"
