+++ app/api/src/app/user/user.module.ts
@@ -1,9 +1,10 @@
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
-import { Module } from "@nestjs/common";
+import { forwardRef, Module } from "@nestjs/common";
+import { AuthModule } from '../../auth/auth.module';
 
 @Module({
-    imports: [],
+    imports: [forwardRef(() => AuthModule)],
     controllers: [UserController],
     providers: [UserService],
 })
