+++ app/api/src/auth/auth.module.ts
@@ -1,14 +1,22 @@
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
-import { Module } from "@nestjs/common";
+import { forwardRef, Module } from "@nestjs/common";
 import { PassportModule } from '@nestjs/passport';
 import { UserModule } from '../app/user/user.module';
 import { FtStrategy } from './strategy/ft.strategy';
 import { AuthRepository } from './repository/auth.repositroy';
+import { JwtModule } from '@nestjs/jwt';
+import { config } from '../config/config';
+import { JwtStrategy } from './strategy/jwt.strategy';
 
 @Module({
-	imports: [PassportModule, UserModule],
+	imports: [PassportModule, forwardRef(() => UserModule),
+		JwtModule.register({
+      secret: config.jwt.secret,
+      signOptions: { expiresIn: config.jwt.expiresIn },
+    }),
+	],
 	controllers: [AuthController],
-	providers: [AuthService, FtStrategy, AuthRepository]
+	providers: [AuthService, FtStrategy, JwtStrategy, AuthRepository]
 })
 export class AuthModule {}
