+++ app/api/src/auth/repository/auth.repositroy.ts
@@ -1,21 +1,23 @@
-import { HttpStatus } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { User } from '@prisma/client';
 import { UserService } from '../../app/user/user.service';
-import { Me } from '../interfaces/auth.interface';
+import { Me } from '../interfaces/intra.interface';
 
+@Injectable()
 export class AuthRepository {
-  constructor(private userService: UserService) {}
+  private userService = new UserService();
 
-  async setupUserAccount(user: Me) : Promise<User> {
-		return await this.userService.CreateUser(this.userService.CreateUserObject(user));
-	}
+  async setupUserAccount(user: Me): Promise<User> {
+    return await this.userService.CreateUser(
+      this.userService.CreateUserObject(user),
+    );
+  }
 
   async userExists(login: string): Promise<boolean> {
-		return await this.userService.getUser(login) ? true : false;
-	}
+    return (await this.userService.getUser(login)) ? true : false;
+  }
 
   async getUser(login: string): Promise<User> {
-		return await this.userService.getUser(login);
-	}
-
+    return await this.userService.getUser(login);
+  }
 }
