+++ app/api/src/auth/strategy/jwt.strategy.ts
@@ -1,0 +1,20 @@
+import { ExtractJwt, Strategy } from 'passport-jwt';
+import { PassportStrategy } from '@nestjs/passport';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { config } from '../../config/config';
+import { Payload } from '../interfaces/auth.interface';
+
+@Injectable()
+export class JwtStrategy extends PassportStrategy(Strategy) {
+  constructor() {
+    super({
+      jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
+      ignoreExpiration: false,
+      secretOrKey: config.jwt.secret,
+    });
+  }
+
+  async validate(payload: any) : Promise<Payload> {
+    return { id: payload.id, login: payload.login };
+  }
+}
